iio: beamformer: adar1000: Initial driver version
The ADAR1000 is a 4-channel, X and Ku frequency band, beamforming core chip
for phased arrays. This device operates in half-duplex between receive and
transmit mode.

A simple 4-wire SPI controls all of the on-chip registers. In addition, two
address pins allow SPI control of up to four devices on thesame serial
lines.

Signed-off-by: Mircea Caprioru <mircea.caprioru@analog.com>