usb: chipidea: core: modernize ulpi reset gpio handling
The ulpi reset gpio code (which is a patch owned by us) was still using
the legacy gpio API which brings some pain from times to times whem
merging to a new kernel version (ex: this merge a macro was dropped from
upstream). Hence, use the proper GPIO consumer API.

On top of that, correctly handle the pin polarity. The pins is active
low so this means we want to request the GPIO asserted to reset the phy
(hence GPIOD_OUT_HIGH) and then bring the device out of reset by
deaserting the pin (hence passing 0 in gpiod_set_value_cansleep().
Things worked because we "cheated" in DT by saying the pin is active
high and then requesting it in the deasserted state (setting the pin to 0
and effectively assert/reset it), Then asserting it by passing 1 to
gpiod_set_value_cansleep() which in fact was deasserting the pin.

While at it, use fsleep() instead of msleep().

Signed-off-by: Nuno Sá <nuno.sa@analog.com>