iio: adc: adrv9002: add pll configuration debugfs support
Add the debugfs interface to read and update the PLL configuration
parameters.

The PLL parameters are used during device initialization and do not
support on-the-fly adjustment. The debugfs entries just map to
the configuration values stored in the instance to allow readback and
updating the PLL configuration prior to loading a new profile and
re-initializing the device.

Signed-off-by: Brent Kowal <brent.kowal@analog.com>
(cherry picked from commit b041aca6e582431ccbd925e4bbccca4428f0a2b2)